% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GSODRdata.R
\docType{package}
\name{GSODRdata}
\alias{GSODRdata}
\alias{GSODRdata-package}
\title{GSODRdata: Extra Climate Data for the GSODR Package}
\description{
This data-only package was created for distributing data useful to the GSODR
package and used in examples.
}
\author{
Adam H Sparks, Tomislav Hengl and Andrew Nelson
}
\references{
(\url{http://chelsa-climate.org/}) / Karger, D. N., Conrad, O., Bohner, J.,
Kawohl, T., Kreft, H., Soria-Auza, R. W., et al. (2016). Climatologies at
high resolution for the Earth land surface areas. arXiv preprint
arXiv:1607.00217

New, M., Lister, D., Hulme, M. and Makin, I., 2002: A high-resolution data
set of surface climate over global land areas. Climate Research 21:1-25
(\href{http://www.int-res.com/abstracts/cr/v21/n1/p1-25.html}{abstract},
\href{http://www.int-res.com/articles/cr2002/21/c021p001.pdf}{paper})

(\url{http://maps.elie.ucl.ac.be/CCI/viewer/index.php})

(\url{http://www.earthenv.org/cloud}) / Wilson AM, Jetz W (2016).
Remotely Sensed High-Resolution Global Cloud Dynamics for Predicting
Ecosystem and Biodiversity Distributions. PLoS Biol 14(3): e1002415.
}
\seealso{
\code{\link{CHELSA}} Data from CHELSA climate layers for GSOD
station locations.

\code{\link{CRU}} Data from CRU CL2.0 climate elements for GSOD
station locations.

\code{\link{ESACCI}} Data from ESA's CCI-LC snow cover probability
for GSOD station locations.

\code{\link{MODCF}} Data from EarthEnv MODIS cloud fraction data for
GSOD station locations.
}

