% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MODCF.R
\docType{data}
\name{MODCF}
\alias{MODCF}
\title{MODCF
\describe{
\item{STNID}{Unique station ID to link to GSOD data from \code{GSODR}}
\item{MODCF_meanannual}{Mean annual cloud fraction}
\item{MODCF_monthlymean_01}{Mean January cloud fraction}
\item{MODCF_monthlymean_02}{Mean February fraction}
\item{MODCF_monthlymean_03}{Mean March cloud fraction}
\item{MODCF_monthlymean_04}{Mean April cloud fraction}
\item{MODCF_monthlymean_05}{Mean May cloud fraction}
\item{MODCF_monthlymean_06}{Mean June cloud fraction}
\item{MODCF_monthlymean_07}{Mean July cloud fraction}
\item{MODCF_monthlymean_08}{Mean August cloud fraction}
\item{MODCF_monthlymean_09}{Mean September cloud fraction}
\item{MODCF_monthlymean_10}{Mean October cloud fraction}
\item{MODCF_monthlymean_11}{Mean November cloud fraction}
\item{MODCF_monthlymean_12}{Mean December cloud fraction}
\item{MODCF_seasonality_concentration}{Cloud fraction seasonality
concentration}
}}
\format{
An object of class \code{data.frame} with 23927 rows and 15 columns.
}
\source{
EarthEnv MODIS cloud fraction (\url{http://www.earthenv.org/cloud})
}
\usage{
MODCF
}
\description{
MODCF
\describe{
\item{STNID}{Unique station ID to link to GSOD data from \code{GSODR}}
\item{MODCF_meanannual}{Mean annual cloud fraction}
\item{MODCF_monthlymean_01}{Mean January cloud fraction}
\item{MODCF_monthlymean_02}{Mean February fraction}
\item{MODCF_monthlymean_03}{Mean March cloud fraction}
\item{MODCF_monthlymean_04}{Mean April cloud fraction}
\item{MODCF_monthlymean_05}{Mean May cloud fraction}
\item{MODCF_monthlymean_06}{Mean June cloud fraction}
\item{MODCF_monthlymean_07}{Mean July cloud fraction}
\item{MODCF_monthlymean_08}{Mean August cloud fraction}
\item{MODCF_monthlymean_09}{Mean September cloud fraction}
\item{MODCF_monthlymean_10}{Mean October cloud fraction}
\item{MODCF_monthlymean_11}{Mean November cloud fraction}
\item{MODCF_monthlymean_12}{Mean December cloud fraction}
\item{MODCF_seasonality_concentration}{Cloud fraction seasonality
concentration}
}
}
\references{
Wilson AM, Jetz W (2016). Remotely Sensed High-Resolution Global
Cloud Dynamics for Predicting Ecosystem and Biodiversity Distributions. PLoS
Biol 14(3): e1002415.
}
\keyword{datasets}
