% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ESACCI.R
\docType{data}
\name{ESACCI}
\alias{ESACCI}
\title{ESACCI
\describe{
\item{STNID}{Unique station ID to link to GSOD data from \code{GSODR}}
\item{ESACCI_snow_prob_Apr_500m}{Mean April snow cover probability}
\item{ESACCI_snow_prob_Aug_500m}{Mean August snow cover probability}
\item{ESACCI_snow_prob_Dec_500m}{Mean December snow cover probability}
\item{ESACCI_snow_prob_Feb_500m}{Mean February snow cover probability}
\item{ESACCI_snow_prob_Jan_500m}{Mean January snow cover probability}
\item{ESACCI_snow_prob_Jul_500m}{Mean July snow cover probability}
\item{ESACCI_snow_prob_Jun_500m}{Mean June snow cover probability}
\item{ESACCI_snow_prob_Mar_500m}{Mean March snow cover probability}
\item{ESACCI_snow_prob_May_500m}{Mean May snow cover probability}
\item{ESACCI_snow_prob_Nov_500m}{Mean November snow cover probability}
\item{ESACCI_snow_prob_Oct_500m}{Mean October snow cover probability}
\item{ESACCI_snow_prob_Sep_500m}{Mean September snow cover probability}
}}
\format{An object of class \code{data.frame} with 23927 rows and 13 columns.}
\source{
ESA's CCI-LC snow cover probability
(\url{http://maps.elie.ucl.ac.be/CCI/viewer/index.php})
}
\usage{
ESACCI
}
\description{
ESACCI
\describe{
\item{STNID}{Unique station ID to link to GSOD data from \code{GSODR}}
\item{ESACCI_snow_prob_Apr_500m}{Mean April snow cover probability}
\item{ESACCI_snow_prob_Aug_500m}{Mean August snow cover probability}
\item{ESACCI_snow_prob_Dec_500m}{Mean December snow cover probability}
\item{ESACCI_snow_prob_Feb_500m}{Mean February snow cover probability}
\item{ESACCI_snow_prob_Jan_500m}{Mean January snow cover probability}
\item{ESACCI_snow_prob_Jul_500m}{Mean July snow cover probability}
\item{ESACCI_snow_prob_Jun_500m}{Mean June snow cover probability}
\item{ESACCI_snow_prob_Mar_500m}{Mean March snow cover probability}
\item{ESACCI_snow_prob_May_500m}{Mean May snow cover probability}
\item{ESACCI_snow_prob_Nov_500m}{Mean November snow cover probability}
\item{ESACCI_snow_prob_Oct_500m}{Mean October snow cover probability}
\item{ESACCI_snow_prob_Sep_500m}{Mean September snow cover probability}
}
}
\keyword{datasets}
