% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GSODRdata.R
\docType{package}
\name{GSODRdata}
\alias{GSODRdata}
\alias{GSODRdata-package}
\title{GSODRdata: Extra Climate Data for the GSODR Package}
\source{
\url{http://chelsa-climate.org/}
\url{http://maps.elie.ucl.ac.be/CCI/viewer/index.php}
\url{http://www.earthenv.org/cloud}
}
\description{
This data-only package was created for distributing data useful to the GSODR
package and used in examples.
}
\author{
Adam H Sparks, Tomislav Hengl and Andrew Nelson
}
\references{
Karger, D. N., Conrad, O., Bohner, J., Kawohl, T., Kreft, H., Soria-Auza,
R. W., \emph{et al}. (2016) Climatologies at high resolution for the Earth
land surface areas. \emph{arXiv preprint} \bold{arXiv:1607.00217}

New, M., Lister, D., Hulme, M. and Makin, I., (2002) A high-resolution data
set of surface climate over global land areas. \emph{Climate Research}
\bold{21}:1--25

Wilson A.M., Jetz W. (2016) Remotely Sensed High-Resolution Global Cloud
Dynamics for Predicting Ecosystem and Biodiversity Distributions.
\emph{PLoS Biol} \bold{14(3)}: e1002415. doi:10.1371/journal. pbio.1002415
}
\seealso{
\code{\link{CHELSA}} Data from CHELSA climate layers for GSOD
station locations.

\code{\link{CRU}} Data from CRU CL2.0 climate elements for GSOD
station locations.

\code{\link{ESACCI}} Data from ESA's CCI-LC snow cover probability
for GSOD station locations.

\code{\link{MODCF}} Data from EarthEnv MODIS cloud fraction data for
GSOD station locations.
}

